### 2019-05-25 #1476 @danielskatz
👋 @MLHale - some of the references have case problems - you might want to enclose anything that should stay upper case in {}s - when you have made changes, to test them, enter a new comment here with ```@whedon generate pdf```
