### 2019-05-25 #1476 @MLHale
@danielskatz I'll fix the case issues, thanks for pointing that out.
