+++ paper.bib
@@ -1,18 +1,18 @@
 @misc{gazepoint, 
-  title={GP3 HD Eye Tracker 150Hz}, 
+  title={GP3 {HD} {E}ye {T}racker 150{H}z}, 
   url={https://www.gazept.com/product/gp3hd/}, 
   journal={Gazepoint}
 }
 
 @article{olsen2012tobii,
-  title={The Tobii I-VT fixation filter},
+  title={The {T}obii {I-VT} fixation filter},
   author={Olsen, Anneli},
   journal={Tobii Technology},
   year={2012}
 }
 
 @misc{djangochannels, 
-  title={Django Channels}, 
+  title={Django {C}hannels}, 
   url={https://channels.readthedocs.io/en/latest/}, 
   journal={Django Channels - 2.1.7 documentation}
 }
@@ -25,14 +25,14 @@
 }
 
 @misc{esdalmaijer2017, 
-  title={esdalmaijer/PyOpenGaze}, 
+  title={esdalmaijer/{P}y{O}pen{G}aze}, 
   url={https://github.com/esdalmaijer/PyOpenGaze}, 
   journal={GitHub}, 
   author={Esdalmaijer}, year={2017}, month={May}
 }
 
 @article{boettiger2015docker,
-  title={An introduction to Docker for reproducible research},
+  title={An introduction to {D}ocker for reproducible research},
   author={Boettiger, Carl},
   journal={ACM SIGOPS Operating Systems Review},
   volume={49},
@@ -43,20 +43,20 @@
 }
 
 @article{hefley2018multimodal,
-  title={Multimodal Data Fusion and Behavioral Analysis Tooling for Exploring Trust, Trust-propensity, and Phishing Victimization in Online Environments},
+  title={Multimodal {D}ata {F}usion and {B}ehavioral {A}nalysis {T}ooling for {E}xploring Trust, {T}rust-propensity, and {P}hishing {V}ictimization in {O}nline {E}nvironments},
   author={Hefley, Michael and Wethor, Gabrielle and Hale, Matthew L},
   year={2018}
 }
 
 @article{hale2016apriori,
-  title={Apriori Prediction of Phishing Victimization Based on Structural Content Factors},
+  title={Apriori {P}rediction of {P}hishing {V}ictimization {B}ased on {S}tructural {C}ontent {F}actors},
   author={Hale, Matthew L and Walter, Charles and Lin, Jessica and Gamble, Rose F},
   journal={International Journal of Services Computing},
   year={2016}
 }
 
 @inproceedings{hale2015measuring,
-  title={Measuring the Potential for Victimization in Malicious Content},
+  title={Measuring the {P}otential for {V}ictimization in {M}alicious {C}ontent},
   author={Hale, Matthew L and Gamble, R and Hale, John and Haney, M and Lin, Jessica and Walter, Charles},
   booktitle={2015 IEEE International Conference on Web Services},
   pages={305--312},
@@ -74,7 +74,7 @@
 }
 
 @article{kasten2006visual,
-  title={Visual field recovery after vision restoration therapy (VRT) is independent of eye movements: an eye tracker study},
+  title={Visual field recovery after vision restoration therapy ({VRT}) is independent of eye movements: an eye tracker study},
   author={Kasten, Erich and Bunzenthal, Ulrike and Sabel, Bernhard A},
   journal={Behavioural brain research},
   volume={175},
