### 2019-05-29 #1476 @MLHale
@cMadan - I don't know of anyone specifically on GitHub except for @esdalmaijer - who created [PyOpenGaze](https://github.com/esdalmaijer/PyOpenGaze), which is one of the dependencies used in this project.

In the broader academic community - there are many authors who've written manuscripts involving gazepoint devices.

A few include:
Pujitha Mannaru - Department of Electrical and Computer Engineering University of Connecticut Storrs USA

Balakumar Balasingam - Department of Electrical and Computer Engineering University of Connecticut Storrs USA

Krishna Pattipati - Warfighter Human Systems Integration Lab Naval Research Laboratory Washington DC USA

Joseph T. Coyne - Warfighter Human Systems Integration Lab Naval Research Laboratory Washington DC USA

Ciara Sibley - Warfighter Human Systems Integration Lab Naval Research Laboratory Washington DC USA

Gregory Funke - Air Force Research Laboratory, Wright-Patterson AFB, OH

Eric - Greenlee - Texas Tech University, Lubbock, TX

Martha Carter - Miami University, Oxford, OH

Allen Dukes - Air Force Research Laboratory, Wright-Patterson AFB, OH

Julie Iskander - Institute for Intelligent Systems Research and Innovation (IISRI), Deakin University, Australia
