### 2019-06-07 #1476 @cMadan
@esdalmaijer @TobiasRoeddiger @EmotionCognitionLab @RingoHHuang @pollen-robotics @nu-childlab, would you be interested and available in reviewing this software paper for publication at JOSS?
