### 2019-07-28 #1476 @cMadan
@MLHale, how much of your submission *requires* access to the Gazepoint hardware? I'm having trouble finding researchers who have the hardware and are interested in reviewing.
