### 2019-07-29 #1476 @MLHale
@cMadan - Well the library sits on-top of the hardware-level libraries and does require the library to stream data.

That said, the program could probably be tested without the hardware - it just won't do anything, since there will be no hardware/hardware libraries to read data from.

Perhaps someone could do a code review or other means of review on the paper/library without testing it in practice? Or possibly I could conduct tests at the behest of the reviewers and post screenshots / videos of the requested tests?
