### 2019-08-01 #1476 @cMadan
@adswa, would you be available and interested in reviewing this submission? I understand that you don't have the hardware for this submission, but a code review would be much appreciated and I should have another reviewer who does have access to the hardware to test the functionality.

@RingoHHuang, thanks for agreeing to review this submission!
