### 2019-08-02 #1620 @whedon
[REVIEW]: Eyestream: An open WebSocket-based middleware for serializing and streaming eye tracker event data from Gazepoint GP3 HD research hardware
**Submitting author:** @MLHale (<a href="http://orcid.org/0000-0002-8433-2744">Matthew Hale</a>)
**Repository:** <a href="https://github.com/MLHale/eyestream" target ="_blank">https://github.com/MLHale/eyestream</a>
**Version:** 1.1

- [x] **Version:** 1.1

- [x] **Version:** 1.1
