### 2019-08-06 MLHale/gazepoint-websockets#3 @adswa
[JOSS]: Missing DOIs
Hi @MLHale,

I'm following up on the [JOSS review](https://github.com/openjournals/joss-reviews/issues/1620) of your software in the form of issues in your repository. As a heads-up: I'm a bit busy, so I'll probably gradually add issues over the course of the next few (hopefully two) weeks.

With regard to the the `paper.md`, could you make sure to include DOIs to the papers you cite to fulfill this aspect of the review:
> References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?
