### 2019-08-06 MLHale/gazepoint-websockets#4 @adswa
[JOSS]: Report issues/Seeking support
With respect to
>  Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I found guidelines on how to contribute in `CONTRIBUTING.md`, but I did not find information on how to report issues/problems. Could you add this to your `CONTRIBUTING.md` file, or elsewhere in your repository?

Belongs to  [this JOSS review](https://github.com/openjournals/joss-reviews/issues/1620).
