### 2019-08-06 MLHale/gazepoint-websockets#5 @adswa
[JOSS]: Installation instructions
Hi @MLHale,

I'd appreciate assistance on the installation of the dependencies. I'm testing the installation using Debian 10 in a Python 2.7.16 virtual environment.

1) One requirement is `pypiwin32`. Is this a requirement for Windows users only, or would I need to have it installed to get your software to work?
2) I'm running into a couple dependency incompatibilities between the Python packages channels, daphne, asgiref, and asgi_redis when installing the other dependencies you list, for example channels and asgi_redis seem to have conflicting asigref version dependencies.

```sh
╰─➤ pip install -U asgiref
DEPRECATION: Python 2.7 will reach the end of its life on January 1st, 2020. Please upgrade your Python as Python 2.7 won't be maintained after that date. A future version of pip will drop support for Python 2.7. More details about Python 2 support in pip, can be found at https://pip.pypa.io/en/latest/development/release-process/#python-2-support
Collecting asgiref
  Using cached https://files.pythonhosted.org/packages/4c/b9/9eb9762c9b43754d49e6b85625c1a5a45673a3083c742be00d8721839b01/asgiref-3.2.1-py2.py3-none-any.whl
ERROR: channels 2.1.3 has requirement asgiref~=2.3, but you'll have asgiref 3.2.1 which is incompatible.
ERROR: asgi-redis 1.4.3 has requirement asgiref~=1.1.2, but you'll have asgiref 3.2.1 which is incompatible.

```

Could you provide me with the version numbers of the dependencies I would need to install? The output of a ``pip freeze`` would be helpful for a start.

Belongs to  [this JOSS review](https://github.com/openjournals/joss-reviews/issues/1620).
