### 2019-08-06 MLHale/gazepoint-websockets#5 @adswa
Ah I see, my bad -- I'm checking the system requirements of the GP3 Analysis software, and saw that it only supports Windows. The installation difficulties I'm experiencing could thus be related to my operating system. @RingoHHuang, you are the other reviewer of this submission, could you check whether the installation of the dependencies works smoother for you, or whether you run into similar dependency conflicts? Please also feel free to close this issue afterwards. Thanks a lot! :slightly_smiling_face:
