### 2019-08-07 MLHale/gazepoint-websockets#5 @RingoHHuang
I'm on Windows 10 Home Edition using Python 3.7.4.

I ran into a bit of a hiccup installing asgi_redis. It seems to replace an existing version of asgiref with a version that is incompatible with some of my other Python packages:

```
ERROR: daphne 2.3.0 has requirement asgiref~=3.0, but you'll have asgiref 1.1.2 which is incompatible.
ERROR: channels 2.2.0 has requirement asgiref~=3.0, but you'll have asgiref 1.1.2 which is incompatible.
Installing collected packages: asgiref, asgi-redis
  Found existing installation: asgiref 3.2.1
    Uninstalling asgiref-3.2.1:
      Successfully uninstalled asgiref-3.2.1
Successfully installed asgi-redis-1.4.3 asgiref-1.1.2
```
