### 2019-08-07 MLHale/gazepoint-websockets#5 @MLHale
(mis-clicked close issue from my phone, sorry about that)
Hi @adswa and @RingoHHuang,

@adswa - Yes GP3, unfortunately only works for windows-based platforms. I will gather specific dependency version numbers and add them in this thread. It might be also more useful if we built a build file. My initial intent was to dockerize everything, but unfortunately the python portions of the app need to be installed natively, due to the GP3 restriction to windows and docker not providing an efficient mapping to screenspace objects (like screen resolution the ilk) that the GP3 needs access to.

@RingoHHuang - Eyestream is written for Python 2.xx (in our install instructions, we suggest 2.7.9, but it should be forwards compatible with any python 2 release). That probably explains the issue you mentioned regarding asgiref. Supporting Python 3.xx should be theoretically possible, but would require a re-write of the underlying communication mechanisms written in Django Channels (which is generating the error message you are getting) and would require us to re-write some of the dependencies we are using.
