### 2019-08-07 MLHale/gazepoint-websockets#4 @MLHale
I will add it this evening, as soon as I am back to my desk. Thanks.
