### 2019-08-07 MLHale/gazepoint-websockets#3 @MLHale
I will update them accordingly this evening when I get back to my desk. Thanks
