+++ paper.bib
@@ -39,13 +39,15 @@
   number={1},
   pages={71--79},
   year={2015},
-  publisher={ACM}
+  publisher={ACM},
+  doi={10.1145/2723872.2723882},
 }
 
 @article{hefley2018multimodal,
   title={Multimodal {D}ata {F}usion and {B}ehavioral {A}nalysis {T}ooling for {E}xploring Trust, {T}rust-propensity, and {P}hishing {V}ictimization in {O}nline {E}nvironments},
   author={Hefley, Michael and Wethor, Gabrielle and Hale, Matthew L},
-  year={2018}
+  year={2018},
+  doi={10.24251/HICSS.2018.108},
 }
 
 @article{hale2016apriori,
@@ -61,7 +63,8 @@
   booktitle={2015 IEEE International Conference on Web Services},
   pages={305--312},
   year={2015},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/ICWS.2015.49},
 }
 
 @inproceedings{hale2015cyberphishing,
@@ -70,7 +73,8 @@
   booktitle={2015 48th Hawaii International Conference on System Sciences},
   pages={5260--5269},
   year={2015},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/HICSS.2015.670},
 }
 
 @article{kasten2006visual,
@@ -81,7 +85,8 @@
   number={1},
   pages={18--26},
   year={2006},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.bbr.2006.07.024},
 }
 
 @inproceedings{goldberg2002eye,
@@ -90,7 +95,8 @@
   booktitle={Proceedings of the 2002 symposium on Eye tracking research \& applications},
   pages={51--58},
   year={2002},
-  organization={ACM}
+  organization={ACM},
+  doi={10.1145/507079.507082},
 }
 
 @article{corcoran2012real,
@@ -101,5 +107,6 @@
   number={2},
   pages={347--355},
   year={2012},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TCE.2012.6227433},
 }
