### 2019-08-07 MLHale/gazepoint-websockets#3 @MLHale
Hi @adswa

Thanks for this note, I have gone through the .bib file and added DOIs (if they exist) for all references.

See 05f8cad linked via commit message above.
