### 2019-08-07 3dab6c7
added information regarding reporting issues and seeking support, addresses #4
