+++ CONTRIBUTING.md
@@ -1,10 +1,10 @@
-# Prior to Contributing
-Please do not reinvent the wheel. Check the set of issues and pull requests to see if your need is already being addressed.
-
 # Contributions
-When contributing to this project please raise your request as an issue, fork the project, and make a series of changes comensurate with your needs.
+When contributing to this project please raise your request as an issue, fork the project, and make a series of changes commensurate with your needs.
 
 Please note we have a code of conduct, please follow it in all your interactions with the project.
+
+## Prior to Contributing
+Please do not reinvent the wheel. Check the set of issues and pull requests to see if your need is already being addressed.
 
 ## Code norms
 Please attempt to follow the norms and styles in the existing codebase. Commented code is welcome.
@@ -13,3 +13,24 @@ Please attempt to follow the norms and styles in the existing codebase. Commente
 When pull requesting your changes please tag the issue you are addressing. Please also ensure any unnecessary install, build, or IDE dependencies are removed before submitting.
 
 Make sure to update the README.md with details of changes, if applicable.
+
+# Reporting issues / problems
+If you have a problem using eyestream, please first check the existing issues and pending pull requests on github by visiting [https://github.com/MLHale/eyestream/issues](https://github.com/MLHale/eyestream/issues). 
+
+If the problem you are having is not listed, please create a new issues. 
+
+## How to file a bug
+1. Go to our [issue tracker on Github](https://github.com/MLHale/eyestream/issues)
+2. Click the green `new issue` button.
+3. Provide any useful information you deem relevant to problem you've identified. At a minimum, please include all of the information below.
+4. Submit your issue. 
+5. (optional) Propose a change using a pull request.
+
+### Information to include when filing a bug 
+* **Descriptive title** - please describe your issue succinctly 
+* **Steps to replicate** - please describe the series of steps that caused your error/problem. Where possible these steps should be repeatable.
+* **Problem description** - please provide specific details about the problem you encounter when you perform the steps.
+* **Operating system and version** - please tell us what operating system you are using and its version as well as versions for any dependencies
+
+# Seeking support
+If you need help using Eyestream, please contact the maintainers of this project, currently [mlhale@unomaha.edu](mlhale@unomaha.edu) or open an issue with a title starting with **(support request)**. We will do our best to respond in a timely manner to help you.
