### 2019-08-07 MLHale/gazepoint-websockets#4 @MLHale
@adswa

I added the requested report/support details patterned after other open source projects. The `contributing.md` file has been updated in 3dab6c7

Thanks.
