### 2019-08-07 879c518
notes regarding windows and python 2, addressing issue #5
