### 2019-08-07 MLHale/gazepoint-websockets#5 @MLHale
I've added some notes to the readme file in 879c518 to clarify that the GP3 only supports windows 8+ and that eyestream is written for Python 2 environments.

I will double check the dependency versions and add those as well.
