### 2019-08-07 MLHale/gazepoint-websockets#3 @adswa
Thanks for getting back to this swiftly!
