### 2019-08-08 MLHale/gazepoint-websockets#5 @RingoHHuang
Hi @MLHale,

Thanks for the clarification! My mistake - didn't see that it was for Python 2. I'm starting the installation steps from the beginning using Python 2.7.9 (Windows 10 Home). I've run into two issues:

- `pip install channels` requires the `pathlib` module, but it doesn't ship with Python 2.7.9. Perhaps include `pip install pathlib` in the installation instructions prior to installing `channels`?
-  Does `asgi_redis` and `channels` depend on different versions of `asgiref`? After `pip install asgi_redis`, I get:
```
ERROR: channels 2.1.3 has requirement asgiref~=2.3, but you'll have asgiref 1.1.2 which is incompatible.
ERROR: daphne 2.3.0 has requirement asgiref~=3.0, but you'll have asgiref 1.1.2 which is incompatible.
```
