### 2019-08-08 d80a3b2
modify readme to reflect change in django channels dependency version, targets issue #5
