### 2019-08-08 MLHale/gazepoint-websockets#5 @MLHale
Added the details I mentioned in commit d80a3b2

Also to answer @adswa 's comment, here is the `pip freeze` results after a fresh install
```
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.24.0
attrs==19.1.0
autobahn==19.8.1
Automat==0.7.0
cffi==1.12.3
channels==1.1.8
constantly==15.1.0
cryptography==2.7
daphne==1.4.2
Django==1.11.23
enum34==1.1.6
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipaddress==1.0.22
msgpack-python==0.5.6
pycparser==2.19
PyHamcrest==1.9.0
pypiwin32==223
pytz==2019.2
pywin32==224
redis==2.10.6
six==1.12.0
Twisted==19.7.0
txaio==18.8.1
zope.interface==4.6.0
```
