### 2019-08-09 MLHale/gazepoint-websockets#5 @RingoHHuang
Thanks @MLHale. I have now installed all the Python packages without any issues. Can you confirm the output below from `docker pull redis` looks fine?
```
Using default tag: latest
latest: Pulling from library/redis
Digest: sha256:854715f5cd1b64d2f62ec219a7b7baceae149453e4d29a8f72cecbb5ac51c4ad
Status: Image is up to date for redis:latest
docker.io/library/redis:latest
```
