### 2019-08-10 MLHale/gazepoint-websockets#5 @adswa
Great, thanks @MLHale. Will close this issue as it is resolved.
