### 2019-08-16 MLHale/gazepoint-websockets#6 @adswa
[JOSS]: (Automated) testing
Hey @MLHale,

as I don't have access to the hardware, I'm trying to figure out how much of the software I can review. Referring to

>  Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?

I wanted to ask whether there is any testing suite. I can't find tests in `api/tests.py`, or in other places in the repo. Am I missing something? Thanks!

Belongs to [this review](https://github.com/openjournals/joss-reviews/issues/1620).
