### 2019-08-17 MLHale/gazepoint-websockets#6 @MLHale
hi @adswa

There are no automated tests in eyestream, but there are manual steps for verifying the functionality of the software.

The manual steps are in the README under the "starting a clientside websocket" heading.

Really there are only 3 user-level tests:

- server starts when called upon to do so via a websocket open message. Expected starting behavior includes launching the gazepoint hardware controller, turning on the gazepoint device, and beginning to stream captured eye data to the websocket which instantiated the channel.
- server streams data captured by the hardware to the websocket client as long as the websocket is open
- server ends a session when called upon to do so via a websocket close event or if the websocket times out. Expected termination behavior is to disable further data collection, gracefully shut down the eye tracker hardware and controller, join the threads supporting the data capture process, and end server processing related to the websocket.

The main processing of these behaviors occurs in the socketconsumers.py file: https://github.com/MLHale/eyestream/blob/master/api/socketconsumers.py

The driver modifications to the PyOpenGaze library that act as middleware between the hardware and server layers are in:

https://github.com/MLHale/eyestream/blob/master/GazepointAnalysis/GazepointWebsockets.py

If needed, I can add these more concrete test criteria to the readme and include some expected outcomes.
