### 2019-08-17 MLHale/eyestream#6 @MLHale
Most of the other files in the repo are basic Django server files that are similar to those of any other django 1.11 installation for an arbitrary web server.
