### 2019-08-26 3623dc4
updated readme to reflect manual testing process, address #6
