### 2019-08-26 MLHale/gazepoint-websockets#6 @MLHale
Hi @adswa - I went ahead and made some changes to the readme file. I recognize that it is not as verbose or complete as an automated test suite, but it does at least provide some manual testing steps and expected outcomes.
