### 2019-09-05 MLHale/gazepoint-websockets#6 @MLHale
Hi @adswa - Is any other action needed on this issue?
