### 2019-09-05 MLHale/gazepoint-websockets#6 @adswa
So sorry, @MLHale , I was sure I'd replied and closed this after your readme update already! Thanks for the poke, and sorry for the delay!
