### 2019-09-05 MLHale/gazepoint-websockets#6 @MLHale
@adswa - No worries! Thanks.
