### 2019-09-07 8a90399
Removed unused dependency apps - addressing #7
