+++ cybertrustgazepoint/settings.py
@@ -39,10 +39,10 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
 
     #app
-    'rest_framework',
+    #'rest_framework',
     'api',
-    'rest_framework_json_api',
-    'corsheaders',
+    #'rest_framework_json_api',
+    #'corsheaders',
 
     #web sockets
     'channels'
