### 2019-09-07 ae6a999
added django picklefield pip install to address #7
