### 2019-09-07 ab97882
further dependency adjustments
