### 2019-09-07 25735dd
further dependency adjustments
