+++ cybertrustgazepoint/settings.py
@@ -39,9 +39,7 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
 
     #app
-    #'rest_framework',
     'api',
-    #'rest_framework_json_api',
     'corsheaders',
 
     #web sockets
