+++ cybertrustgazepoint/urls.py
@@ -17,7 +17,6 @@ from django.contrib import admin
 
 from django.conf.urls import *
 
-#Django Rest Framework
 from django.views.decorators.csrf import csrf_exempt
 
 urlpatterns = [
