### 2019-09-09 MLHale/gazepoint-websockets#8 @adswa
[JOSS] Minor paper remarks
Thanks for the clear, concise paper with useful illustrations. I only have very minor comments about formalities in ``paper.md``:
- missing space in paragraph 2, sentence 1, after "While the Gazepoint GP3 HD" and the citation.
- I'm not a native English speaker, so please disregard if I am wrong here: I believe the proper capitalization of the subheading "On-going research projects using Eyestream" would be "On-Going Research Projects Using Eyestream", and similarly, the title would be "Eyestream: An Open Websocket-Based Middleware for Serializing and Streaming Eye Tracker Event Data from Gazepoint GP3 HD Research Hardware"
- Currently, there is inconsistent capitalization in figure captions, some captions start with a ``>`` and some not, and Figure 5's caption is starting in the same line as the figure instead of underneath it
- The array of citations in the "On-going research" section would probably benefit from combining them into a single bracket, with individual citations separated by semicolons (e.g. ``Blah blah [@smith04; @doe99].``, see [this reference](https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax))

belongs to [this review](https://github.com/openjournals/joss-reviews/issues/1620)
