### 2019-09-09 MLHale/gazepoint-websockets#8 @MLHale
Thanks @adswa - I will make these corrections asap (likely later today). Thanks again for your efforts and time in this review.
