### 2019-09-10 #1620 @adswa
Hey everyone, from my POV and based on what I am able to review given I don't have the hardware and OS available to test the full functionality, my review is finished.
I can confirm that the items on the checklist apply with the exception of "functionality". I have a minor issue open regarding the paper, but once this is addressed, I'd be happy to approve this submission - all with the explicit notion that I was not able to actually test the functionality. Thanks to @MLHale for providing an open source interface to make the raw data accessible, and for being so swift and diligent in replying to my issues, and thanks to everyone involved for this wonderful way of publishing and reviewing research items!
