### 2019-09-10 59308af
made suggested changes as @adwsa suggested in #8 to correct capitalization and citation and figure small issues
