+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'Eyestream: An open WebSocket-based middleware for serializing and streaming eye tracker event data from Gazepoint GP3 HD research hardware'
+title: 'Eyestream: An Open WebSocket-based Middleware for Serializing and Streaming Eye Tracker Event Data from Gazepoint GP3 HD Research Hardware'
 tags:
   - Python
   - eye tracker
@@ -21,10 +21,10 @@ bibliography: paper.bib
 # Summary
 Eye trackers enable a wide range of research and development activities in academia, medicine, and industry. Researchers and developers using eye trackers are able to collect rich user interaction data such as user fixation durations, points of gaze, and 3D models of user pupils that enable unique analysis and applications in areas such as user-interaction design, rehabilitative/restorative medicine, and marketing. Most eye trackers, including the two with the largest market share `Gazepoint GP3 HD`[@gazepoint] and the `Tobii Pro X3-120` [@olsen2012tobii], ship with proprietary desktop software packages that collect, house, and analyze captured data. This closed source delivery model makes it difficult for researchers and developers to directly interact with raw eye tracker data, which prevents custom analysis and hinders application development and integration efforts. Further, most closed-source eye tracker analysis tools do not support streaming and real-time analysis applications.
 
-While the Gazepoint GP3 HD[@gazepoint] is designed to work with Gazepoint’s proprietary analysis software, it also provides an underutilized alternative low-level XML-based API called OpenGaze for extracting data from the hardware. We have created a tool on top of OpenGaze called `Eyestream`. `Eyestream` is capable of serializing hardware-captured eye tracker data as JSON and streaming it in real-time, at frequencies up to 150hz, to desktop, web, or mobile applications. `Eyestream` is an open source package designed to operate as middleware between the Gazepoint GP3 and applications that consume eye tracker data for analysis or visualization purposes. It was implemented using Python 2.7, a web socket server framework called `Django Channels`[@djangochannels], an in-memory database caching tool called `Redis`[@carlson2013redis], and a modified version of `PyOpenGaze`[@esdalmaijer2017]. `Eyestream` is containerized using `Docker` [@boettiger2015docker]. 
+While the Gazepoint GP3 HD [@gazepoint] is designed to work with Gazepoint’s proprietary analysis software, it also provides an underutilized alternative low-level XML-based API called OpenGaze for extracting data from the hardware. We have created a tool on top of OpenGaze called `Eyestream`. `Eyestream` is capable of serializing hardware-captured eye tracker data as JSON and streaming it in real-time, at frequencies up to 150hz, to desktop, web, or mobile applications. `Eyestream` is an open source package designed to operate as middleware between the Gazepoint GP3 and applications that consume eye tracker data for analysis or visualization purposes. It was implemented using Python 2.7, a web socket server framework called `Django Channels`[@djangochannels], an in-memory database caching tool called `Redis`[@carlson2013redis], and a modified version of `PyOpenGaze`[@esdalmaijer2017]. `Eyestream` is containerized using `Docker` [@boettiger2015docker]. 
 
-![Figure 1: System-level Architecture of Eyestream](assets/paper-040a5e2f.png)
-> Figure 1: System-level Architecture of Eyestream
+![Figure 1: System-level architecture of eyestream](assets/paper-040a5e2f.png)
+> Figure 1: System-level architecture of eyestream
  
 # Eyestream and Its Uses
[…]
