### 2019-09-10 MLHale/gazepoint-websockets#8 @MLHale
Hi @adswa - Sorry I didn't get to this last night. Thanks again so much for all of your inputs in the review. I quite like this journal review format - it feels more collaborative and constructive than typical academic review processes.
