### 2019-09-10 #1620 @MLHale
@adswa - I have implemented your change list. Really thanks again for all of your review inputs and thoughtful contributions. I also like this collaborative process - it was been very constructive to receive and implement feedback in this way!
