### 2019-09-13 0d1eba5
clarified run instructions for gazepoint testing - this clarification relates to issue #10
