### 2019-09-16 #1620 @RingoHHuang
Hi @cMadan @MLHale @adswa,

I've just tested Eyestream using my Gazepoint 60 Hz eye-tracker and can confirm its functionality. My review of this submission is complete, and I'm ready to approve it. Thanks @MLHale for your help in addressing my many installation and testing questions, and for amending the documentation accordingly. I think Eyestream will be a valuable addition to the eye-tracking development community and will broaden the use case for Gazepoint eye-trackers. I'm excited to see future applications built off of this software. Thanks to all involved - it was a pleasure reviewing this submission!
