### 2019-10-18 #1620 @cMadan
@MLHale, it looks like the bib needs a few fixes (missing year, missing capitalisation, missing page numbers).
