### 2019-11-02 7a95b74
updates to bib per editor comments
