+++ paper.bib
@@ -1,38 +1,41 @@
 @misc{gazepoint, 
   title={GP3 {HD} {E}ye {T}racker 150{H}z}, 
   url={https://www.gazept.com/product/gp3hd/}, 
-  journal={Gazepoint}
+  journal={Gazepoint},
+  year={2018},
 }
 
 @article{olsen2012tobii,
   title={The {T}obii {I-VT} fixation filter},
   author={Olsen, Anneli},
   journal={Tobii Technology},
-  year={2012}
+  year={2012},
+  pages={1--21},
 }
 
 @misc{djangochannels, 
   title={Django {C}hannels}, 
   url={https://channels.readthedocs.io/en/latest/}, 
   journal={Django Channels - 2.1.7 documentation}
+  year={2018},
 }
 
 @book{carlson2013redis,
-  title={Redis in action},
+  title={Redis in {A}ction},
   author={Carlson, Josiah L},
   year={2013},
   publisher={Manning Publications Co.}
 }
 
 @misc{esdalmaijer2017, 
-  title={esdalmaijer/{P}y{O}pen{G}aze}, 
+  title={Esdalmaijer/{P}y{O}pen{G}aze}, 
   url={https://github.com/esdalmaijer/PyOpenGaze}, 
   journal={GitHub}, 
   author={Esdalmaijer}, year={2017}, month={May}
 }
 
 @article{boettiger2015docker,
-  title={An introduction to {D}ocker for reproducible research},
+  title={An {I}ntroduction to {D}ocker for {R}eproducible {R}esearch},
   author={Boettiger, Carl},
   journal={ACM SIGOPS Operating Systems Review},
   volume={49},
@@ -47,6 +50,7 @@
   title={Multimodal {D}ata {F}usion and {B}ehavioral {A}nalysis {T}ooling for {E}xploring Trust, {T}rust-propensity, and {P}hishing {V}ictimization in {O}nline {E}nvironments},
   author={Hefley, Michael and Wethor, Gabrielle and Hale, Matthew L},
   year={2018},
+  pages={862--871},
   doi={10.24251/HICSS.2018.108},
 }
 
@@ -54,6 +58,7 @@
   title={Apriori {P}rediction of {P}hishing {V}ictimization {B}ased on {S}tructural {C}ontent {F}actors},
   author={Hale, Matthew L and Walter, Charles and Lin, Jessica and Gamble, Rose F},
   journal={International Journal of Services Computing},
+  pages={1--13},
   year={2016}
 }
 
@@ -68,7 +73,7 @@
 }
 
 @inproceedings{hale2015cyberphishing,
-  title={CyberPhishing: a game-based platform for phishing awareness testing},
+  title={Cyber{P}hishing: {A} {G}ame-based {P}latform for {P}hishing {A}wareness {T}esting},
   author={Hale, Matthew L and Gamble, Rose F and Gamble, Philip},
   booktitle={2015 48th Hawaii International Conference on System Sciences},
   pages={5260--5269},
@@ -78,7 +83,7 @@
 }
 
 @article{kasten2006visual,
-  title={Visual field recovery after vision restoration therapy ({VRT}) is independent of eye movements: an eye tracker study},
+  title={Visual {F}ield {R}ecovery {A}fter {V}ision {R}estoration {T}herapy ({VRT}) is {I}ndependent of {E}ye {M}ovements: {A}n {E}ye {T}racker {S}tudy},
   author={Kasten, Erich and Bunzenthal, Ulrike and Sabel, Bernhard A},
   journal={Behavioural brain research},
   volume={175},
@@ -90,7 +95,7 @@
 }
 
 @inproceedings{goldberg2002eye,
-  title={Eye tracking in web search tasks: design implications},
[…]
