### 2019-11-02 #1620 @MLHale
@danielskatz @cMadan - Yes, just now - my apologies. Thanks for the reminder. I apparently had an un-pushed commit.

Changes are pushed now on commit:
https://github.com/MLHale/eyestream/commit/7a95b7423f7ad2a4f3d1053025b890baa189f42d
