+++ paper.bib
@@ -16,7 +16,7 @@
 @misc{djangochannels, 
   title={Django {C}hannels}, 
   url={https://channels.readthedocs.io/en/latest/}, 
-  journal={Django Channels - 2.1.7 documentation}
+  journal={Django Channels - 2.1.7 documentation},
   year={2018},
 }
 
