### 2019-11-05 #1620 @cMadan
@MLHale, can you fix the author name for the pyOpenGaze reference so it's the person's actual name (Edwin Dalmaijer)? Also add the DOI as suggested above, as it does seem correct (http://hipore.com/stsc/contents.html#v5.n1.2017).

After that we can move on to the last stages before acceptance.
