### 2019-11-13 2f6706f
fix author name and doi as suggested by https://github.com/openjournals/joss-reviews/issues/1620#issuecomment-549704376
