+++ paper.bib
@@ -31,7 +31,7 @@
   title={Esdalmaijer/{P}y{O}pen{G}aze}, 
   url={https://github.com/esdalmaijer/PyOpenGaze}, 
   journal={GitHub}, 
-  author={Esdalmaijer}, year={2017}, month={May}
+  author={Dalmaijer, Edwin}, year={2017}, month={May}
 }
 
 @article{boettiger2015docker,
@@ -59,7 +59,8 @@
   author={Hale, Matthew L and Walter, Charles and Lin, Jessica and Gamble, Rose F},
   journal={International Journal of Services Computing},
   pages={1--13},
-  year={2016}
+  year={2016},
+  doi={https://doi.org/10.29268/stsc.2017.5.1.1}
 }
 
 @inproceedings{hale2015measuring,
