### 2019-11-18 #1620 @cMadan
@MLHale, everything looks good to me!

To move forward with accepting your submission, there are a few last things to take care of:

You may find this helpful: https://guides.github.com/activities/citable-code/
