### 2019-11-20 #1620 @MLHale
Hi @cMadan - Thanks!

Here is the tagged release
https://github.com/MLHale/eyestream/releases/tag/1.1
