### 2019-11-20 #1620 @MLHale
The Zenodo archive is available here:
10.5281/zenodo.3549022

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3549022.svg)](https://doi.org/10.5281/zenodo.3549022)
