### 2019-11-20 #1620 @cMadan
@openjournals/joss-eics, I think we're all set to accept here!
