### 2019-11-20 #1620 @MLHale
@cMadan - Thanks again for all of your help. I want to reiterate how awesome this review process has been. Seriously the most transparent, useful, and enjoyable review process I've seen.

I really like that reviewer feedback is so collaborative, can be incorporated into the work in near real-time, and then iterated on if the fix didn't address the feedback. This kind of process makes for a better product and is (I hope) the future of academia and publishing!

I'll definitely be publishing in this venue again!
