### 2019-11-21 37652d0
Some typos fixed
mostly spaces before a reference
