### 2019-11-21 #1620 @kthyng
Hi @MLHale I've made some tweaks to your paper.md and paper.bib to improve spacing around the references and preserve capitalization (mostly). Please merge PRs [11](https://github.com/MLHale/eyestream/pull/11) and [12](https://github.com/MLHale/eyestream/pull/12) if you please.
