### 2019-11-21 #1620 @MLHale
@kthyng Thanks for these improvements. I just accepted the PRs.
