### 2019-11-21 #1620 @kthyng
Congrats to @MLHale on your new paper! Thanks to editor @cMadan and reviewers @RingoHHuang and @adswa — we wouldn't be here without you!
